src/requirements.txt
PyBambooHR==0.8.1
google-api-python-client==2.47.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1